fix todo by moving ShouldConsiderSelfKeyInConsensus to common package
