+++ process/aotSelection/aotSelector_test.go
@@ -391,41 +391,6 @@ func TestAOTSelector_IsSelfLeaderForRoundNoKeyMatch(t *testing.T) {
 	assert.False(t, sel.isSelfLeaderForRound([]byte("rand"), 10, 1))
 }
 
-func TestAOTSelector_ShouldConsiderSelfKeyInConsensusMainMachine(t *testing.T) {
-	t.Parallel()
-
-	args := createDefaultArgs()
-	args.NodeRedundancy = &consensusMock.NodeRedundancyHandlerStub{
-		IsRedundancyNodeCalled: func() bool { return false },
-	}
-	sel, _ := NewAOTSelector(args)
-	assert.True(t, sel.shouldConsiderSelfKeyInConsensus())
-}
-
-func TestAOTSelector_ShouldConsiderSelfKeyInConsensusRedundancyNodeMainActive(t *testing.T) {
-	t.Parallel()
-
-	args := createDefaultArgs()
-	args.NodeRedundancy = &consensusMock.NodeRedundancyHandlerStub{
-		IsRedundancyNodeCalled:    func() bool { return true },
-		IsMainMachineActiveCalled: func() bool { return true },
-	}
-	sel, _ := NewAOTSelector(args)
-	assert.False(t, sel.shouldConsiderSelfKeyInConsensus())
-}
-
-func TestAOTSelector_ShouldConsiderSelfKeyInConsensusRedundancyNodeMainInactive(t *testing.T) {
-	t.Parallel()
-
-	args := createDefaultArgs()
-	args.NodeRedundancy = &consensusMock.NodeRedundancyHandlerStub{
-		IsRedundancyNodeCalled:    func() bool { return true },
-		IsMainMachineActiveCalled: func() bool { return false },
-	}
-	sel, _ := NewAOTSelector(args)
-	assert.True(t, sel.shouldConsiderSelfKeyInConsensus())
-}
-
 func TestAOTSelector_GetPreSelectedTransactionsCacheMiss(t *testing.T) {
 	t.Parallel()
